[chore/#59] fetchPlaceNearby 함수 작성 파일 변경
